nothing depends on fplll.h
fplll.h includes everything so anything that includes it is recompiled
whenever anything changes, which can be slow.

See #160
