+++ fplll/fplll.h
@@ -29,46 +29,4 @@
 #include "wrapper.h"
 #include "pruner.h"
 
-FPLLL_BEGIN_NAMESPACE
-
-#define FPLLL_DECLARE_LLL(T)                                                \
-int lll_reduction(ZZ_mat<T>& b,                                              \
-        double delta = LLL_DEF_DELTA, double eta = LLL_DEF_ETA,             \
-        LLLMethod method = LM_WRAPPER, FloatType floatType = FT_DEFAULT,    \
-        int precision = 0, int flags = LLL_DEFAULT);                        \
-                                                                            \
-int lll_reduction(ZZ_mat<T>& b, ZZ_mat<T>& u,                                \
-        double delta = LLL_DEF_DELTA, double eta = LLL_DEF_ETA,             \
-        LLLMethod method = LM_WRAPPER, FloatType floatType = FT_DEFAULT,    \
-        int precision = 0, int flags = LLL_DEFAULT);                        \
-                                                                            \
-int lll_reduction(ZZ_mat<T>& b, ZZ_mat<T>& u, ZZ_mat<T>& u_inv,               \
-        double delta = LLL_DEF_DELTA, double eta = LLL_DEF_ETA,             \
-        LLLMethod method = LM_WRAPPER, FloatType floatType = FT_DEFAULT,    \
-        int precision = 0, int flags = LLL_DEFAULT);
-
-FPLLL_DECLARE_LLL(mpz_t)
-
-#ifdef FPLLL_WITH_ZLONG
-FPLLL_DECLARE_LLL(long)
-#endif
-
-#ifdef FPLLL_WITH_ZDOUBLE
-FPLLL_DECLARE_LLL(double)
-#endif
-
-
-int bkz_reduction(IntMatrix* B, IntMatrix* U, const BKZParam& param, FloatType float_type=FT_DEFAULT, int precision=0);
-int bkz_reduction(IntMatrix& b, int block_size, int flags = BKZ_DEFAULT, FloatType float_type=FT_DEFAULT, int precision=0);
-int bkz_reduction(IntMatrix& b, IntMatrix& u, int block_size, int flags = BKZ_DEFAULT, FloatType float_type=FT_DEFAULT, int precision=0);
-
-int hkz_reduction(IntMatrix& b, int flags = HKZ_DEFAULT, FloatType float_type=FT_DEFAULT, int precision=0);
-
-/**
- * Returns the string corresponding to an error code of LLL/BKZ.
- */
-const char* get_red_status_str(int status);
-
-FPLLL_END_NAMESPACE
-
 #endif
