+++ fplll/sieve/sampler_basic.h
@@ -4,9 +4,10 @@
 #ifndef FPLLL_SIEVE_SAMPLER_BASIC_H
 #define FPLLL_SIEVE_SAMPLER_BASIC_H
 
-#include "../fplll.h"
-#include "../gso.cpp"
-
+#include "fplll/defs.h"
+#include "fplll/nr/nr_Z.inl"
+#include "fplll/nr/matrix.h"
+#include "fplll/gso.h"
 
 using namespace std;
 using namespace fplll;
